added OfferCard
